requirements.txt
torch
torchvision
pandas
scikit-learn
matplotlib
tqdm
Pillow
